requirements.txt
pandas==2.2.2
numpy>=1.21.0

scikit-learn==1.3.2
imbalanced-learn>=0.10.0
category_encoders==2.6.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.0

ydata-profiling>=4.6.0

nltk>=3.8.0

statsmodels>=0.14.0

joblib>=1.3.0
thefuzz>=0.19.0
gdown>=4.7.0

seaborn>=0.12.0
matplotlib>=3.6.0
plotly>=5.15.0
pyLDAvis>=3.4.0

sentence-transformers>=2.2.0
hdbscan>=0.8.0
torch>=1.13.0

pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
pytest-asyncio>=0.23.0
locust>=2.0.0

featuretools>=1.26.0

pydantic>=2.4.0

psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

azure-storage-blob>=12.19.0
azure-identity>=1.15.0

google-generativeai>=0.5.2

neo4j>=5.12.0

python-dotenv==1.0.1
langchain==0.2.1
langchain-core>=0.2.36,<0.3.0
langchain-google-genai==1.0.5
langchain-ollama>=0.1.0,<0.2.0
langgraph==0.0.60
duckduckgo-search>=5.0.0
e2b-code-interpreter
werkzeug==3.0.3
psutil>=5.9.8